requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.12
djangorestframework==3.5.3
enum34==1.1.6
factory-boy==2.7.0
fake-factory==0.7.2
ipaddress==1.0.17
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
